Fix tokenizers w/ `padding: true` and non-null `max_length` (#1284)
* tokenizers respect padding: true with non-null max_length

This commit changes the behavior of tokenizers to match the
behavior described in the docs and the behavior of the Python
library.

Before this commit, passing
{
  padding: true,
  max_length: 512
}
or
{
  padding: 'max_length',
  max_length: 512
}
would both always pad all outputs to 512 tokens.

After this change,
{
  padding: true,
  max_length: 512
}
will now pad the outputs to match the longest encoding
or max_length, whichever is shorter.

This commit  also adds a test to prevent regressions.

* Revamp tokenizer padding/truncation test suite

* Fix tokenization padding/truncation logic

* nit

---------

Co-authored-by: Joshua Lochner <26504141+xenova@users.noreply.github.com>